arreglo MusicianService, properties y colecciones postman

+++ src/main/java/com/svalero/music/rights/service/MusicianService.java
@@ -127,6 +127,9 @@ public class MusicianService {
     }
 
     public void delete(long id) {
+        if (!musicianRepository.existsById(id)) {
+            throw new MusicianNotFoundException();
+        }
         musicianRepository.deleteById(id);
     }
